Add group intent modules for MLS
