+++ xmtp_mls/src/groups/intents.rs
@@ -2,8 +2,7 @@ use openmls::prelude::{
     tls_codec::{Error as TlsCodecError, Serialize},
     MlsMessageOut,
 };
-use prost::{bytes::Bytes, DecodeError, Message};
-use std::collections::{HashMap, HashSet};
+use prost::{DecodeError, Message};
 use thiserror::Error;
 
 use xmtp_proto::xmtp::mls::database::{
@@ -12,25 +11,11 @@ use xmtp_proto::xmtp::mls::database::{
         Installation as InstallationProto, Kind as PostCommitActionKind,
         SendWelcomes as SendWelcomesProto,
     },
-    send_message_data::{Version as SendMessageVersion, V1 as SendMessageV1},
-    update_admin_lists_data::{Version as UpdateAdminListsVersion, V1 as UpdateAdminListsV1},
-    update_group_membership_data::{
-        Version as UpdateGroupMembershipVersion, V1 as UpdateGroupMembershipV1,
-    },
-    update_metadata_data::{Version as UpdateMetadataVersion, V1 as UpdateMetadataV1},
-    update_permission_data::{self, Version as UpdatePermissionVersion, V1 as UpdatePermissionV1},
     AccountAddresses, AddressesOrInstallationIds as AddressesOrInstallationIdsProtoWrapper,
-    InstallationIds, PostCommitAction as PostCommitActionProto, SendMessageData,
-    UpdateAdminListsData, UpdateGroupMembershipData, UpdateMetadataData, UpdatePermissionData,
+    InstallationIds, PostCommitAction as PostCommitActionProto,
 };
 
-use super::{
-    group_membership::GroupMembership,
-    group_mutable_metadata::MetadataField,
-    group_permissions::{MembershipPolicies, MetadataPolicies, PermissionsPolicies},
-    scoped_client::ScopedGroupClient,
-    GroupError, MlsGroup,
-};
+use super::{scoped_client::ScopedGroupClient, GroupError, MlsGroup};
 use crate::{
     configuration::GROUP_KEY_ROTATION_INTERVAL_NS,
     verified_key_package_v2::{KeyPackageVerificationError, VerifiedKeyPackageV2},
@@ -42,6 +27,20 @@ use xmtp_db::{
     group_intent::{IntentKind, NewGroupIntent, StoredGroupIntent},
     ProviderTransactions,
 };
+
+mod admin;
+mod group_membership;
+mod key_update;
+mod metadata;
+mod permission;
+mod send_message;
+pub use admin::*;
+pub use group_membership::*;
+pub use key_update::*;
+pub use metadata::*;
+pub use permission::*;
+pub use send_message::*;
+
 #[derive(Debug, Error)]
 pub enum IntentError {
     #[error("decode error: {0}")]
@@ -123,42 +122,6 @@ impl<ScopedClient: ScopedGroupClient> MlsGroup<ScopedClient> {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct SendMessageIntentData {
-    pub message: Vec<u8>,
-}
-
-impl SendMessageIntentData {
-    pub fn new(message: Vec<u8>) -> Self {
-        Self { message }
-    }
-
-    pub(crate) fn to_bytes(&self) -> Vec<u8> {
-        SendMessageData {
-            version: Some(SendMessageVersion::V1(SendMessageV1 {
-                payload_bytes: self.message.clone(),
-            })),
-        }
-        .encode_to_vec()
-    }
-
-    pub(crate) fn from_bytes(data: &[u8]) -> Result<Self, IntentError> {
-        let msg = SendMessageData::decode(data)?;
-        let payload_bytes = match msg.version {
-            Some(SendMessageVersion::V1(v1)) => v1.payload_bytes,
-            None => return Err(IntentError::MissingPayload),
-        };
-
-        Ok(Self::new(payload_bytes))
-    }
-}
-
-impl From<SendMessageIntentData> for Vec<u8> {
-    fn from(intent: SendMessageIntentData) -> Self {
-        intent.to_bytes()
-    }
-}
-
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub enum AddressesOrInstallationIds {
     AccountAddresses(Vec<String>),
@@ -215,461 +178,6 @@ impl From<Vec<Address>> for AddressesOrInstallationIds {
 impl From<Vec<Vec<u8>>> for AddressesOrInstallationIds {
     fn from(installation_ids: Vec<Vec<u8>>) -> Self {
         AddressesOrInstallationIds::InstallationIds(installation_ids)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct UpdateMetadataIntentData {
-    pub field_name: String,
-    pub field_value: String,
-}
-
-impl UpdateMetadataIntentData {
-    pub fn new(field_name: String, field_value: String) -> Self {
-        Self {
-            field_name,
-            field_value,
-        }
-    }
-
-    pub fn new_update_group_name(group_name: String) -> Self {
-        Self {
-            field_name: MetadataField::GroupName.to_string(),
-            field_value: group_name,
-        }
-    }
-
-    pub fn new_update_group_image_url_square(group_image_url_square: String) -> Self {
-        Self {
-            field_name: MetadataField::GroupImageUrlSquare.to_string(),
-            field_value: group_image_url_square,
-        }
-    }
-
-    pub fn new_update_group_description(group_description: String) -> Self {
-        Self {
-            field_name: MetadataField::Description.to_string(),
-            field_value: group_description,
-        }
-    }
-
-    pub fn new_update_conversation_message_disappear_from_ns(from_ns: i64) -> Self {
-        Self {
-            field_name: MetadataField::MessageDisappearFromNS.to_string(),
-            field_value: from_ns.to_string(),
-        }
-    }
-    pub fn new_update_conversation_message_disappear_in_ns(in_ns: i64) -> Self {
-        Self {
-            field_name: MetadataField::MessageDisappearInNS.to_string(),
-            field_value: in_ns.to_string(),
-        }
-    }
-
-    pub fn new_update_group_min_version_to_match_self(min_version: String) -> Self {
-        Self {
-            field_name: MetadataField::MinimumSupportedProtocolVersion.to_string(),
-            field_value: min_version,
-        }
-    }
-}
-
-impl From<UpdateMetadataIntentData> for Vec<u8> {
-    fn from(intent: UpdateMetadataIntentData) -> Self {
-        let mut buf = Vec::new();
-
-        UpdateMetadataData {
-            version: Some(UpdateMetadataVersion::V1(UpdateMetadataV1 {
-                field_name: intent.field_name.to_string(),
-                field_value: intent.field_value.clone(),
-            })),
-        }
-        .encode(&mut buf)
-        .expect("encode error");
-
-        buf
-    }
-}
-
-impl TryFrom<Vec<u8>> for UpdateMetadataIntentData {
-    type Error = IntentError;
-
-    fn try_from(data: Vec<u8>) -> Result<Self, Self::Error> {
-        let msg = UpdateMetadataData::decode(Bytes::from(data))?;
-
-        let field_name = match msg.version {
-            Some(UpdateMetadataVersion::V1(ref v1)) => v1.field_name.clone(),
-            None => return Err(IntentError::MissingPayload),
-        };
-        let field_value = match msg.version {
-            Some(UpdateMetadataVersion::V1(ref v1)) => v1.field_value.clone(),
-            None => return Err(IntentError::MissingPayload),
-        };
-
-        Ok(Self::new(field_name, field_value))
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct UpdateGroupMembershipResult {
-    pub added_members: HashMap<String, u64>,
-    pub removed_members: Vec<String>,
-    pub failed_installations: Vec<Vec<u8>>,
-}
-
-impl UpdateGroupMembershipResult {
-    pub fn new(
-        added_members: HashMap<String, u64>,
-        removed_members: Vec<String>,
-        failed_installations: Vec<Vec<u8>>,
-    ) -> Self {
-        Self {
-            added_members,
-            removed_members,
-            failed_installations,
-        }
-    }
-}
-
-impl From<UpdateGroupMembershipIntentData> for UpdateGroupMembershipResult {
-    fn from(value: UpdateGroupMembershipIntentData) -> Self {
-        UpdateGroupMembershipResult::new(
-            value.membership_updates,
-            value.removed_members,
-            value.failed_installations,
-        )
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct UpdateGroupMembershipIntentData {
-    pub membership_updates: HashMap<String, u64>,
-    pub removed_members: Vec<String>,
-    pub failed_installations: Vec<Vec<u8>>,
-}
-
-impl UpdateGroupMembershipIntentData {
-    pub fn new(
-        membership_updates: HashMap<String, u64>,
-        removed_members: Vec<String>,
-        failed_installations: Vec<Vec<u8>>,
-    ) -> Self {
-        Self {
-            membership_updates,
-            removed_members,
-            failed_installations,
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.membership_updates.is_empty()
-            && self.removed_members.is_empty()
-            && self.failed_installations.is_empty()
-    }
-
-    pub fn apply_to_group_membership(&self, group_membership: &GroupMembership) -> GroupMembership {
-        tracing::info!("old group membership: {:?}", group_membership.members);
-        let mut new_membership = group_membership.clone();
-        for (inbox_id, sequence_id) in self.membership_updates.iter() {
-            new_membership.add(inbox_id.clone(), *sequence_id);
-        }
-
-        for inbox_id in self.removed_members.iter() {
-            new_membership.remove(inbox_id)
-        }
-
-        new_membership.failed_installations = new_membership
-            .failed_installations
-            .into_iter()
-            .chain(self.failed_installations.iter().cloned())
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect();
-
-        tracing::info!("updated group membership: {:?}", new_membership.members);
-        new_membership
-    }
-}
-
-impl From<UpdateGroupMembershipIntentData> for Vec<u8> {
-    fn from(intent: UpdateGroupMembershipIntentData) -> Self {
-        let mut buf = Vec::new();
-
-        UpdateGroupMembershipData {
-            version: Some(UpdateGroupMembershipVersion::V1(UpdateGroupMembershipV1 {
-                membership_updates: intent.membership_updates,
-                removed_members: intent.removed_members,
-                failed_installations: intent.failed_installations,
-            })),
-        }
-        .encode(&mut buf)
-        .expect("encode error");
-
-        buf
-    }
-}
-
-impl TryFrom<Vec<u8>> for UpdateGroupMembershipIntentData {
-    type Error = IntentError;
-
-    fn try_from(data: Vec<u8>) -> Result<Self, Self::Error> {
-        if let UpdateGroupMembershipData {
-            version: Some(UpdateGroupMembershipVersion::V1(v1)),
-        } = UpdateGroupMembershipData::decode(data.as_slice())?
-        {
-            Ok(Self::new(
-                v1.membership_updates,
-                v1.removed_members,
-                v1.failed_installations,
-            ))
-        } else {
-            Err(IntentError::MissingPayload)
-        }
-    }
-}
-
-impl<'a> TryFrom<&'a [u8]> for UpdateGroupMembershipIntentData {
-    type Error = IntentError;
-
-    fn try_from(data: &'a [u8]) -> Result<Self, Self::Error> {
-        if let UpdateGroupMembershipData {
-            version: Some(UpdateGroupMembershipVersion::V1(v1)),
-        } = UpdateGroupMembershipData::decode(data)?
-        {
-            Ok(Self::new(
-                v1.membership_updates,
-                v1.removed_members,
-                v1.failed_installations,
-            ))
-        } else {
-            Err(IntentError::MissingPayload)
-        }
-    }
-}
-#[repr(i32)]
-#[derive(Debug, Clone, PartialEq)]
-pub enum AdminListActionType {
-    Add = 1,         // Matches ADD_ADMIN in Protobuf
-    Remove = 2,      // Matches REMOVE_ADMIN in Protobuf
-    AddSuper = 3,    // Matches ADD_SUPER_ADMIN in Protobuf
-    RemoveSuper = 4, // Matches REMOVE_SUPER_ADMIN in Protobuf
-}
-
-impl TryFrom<i32> for AdminListActionType {
-    type Error = IntentError;
-
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
-        match value {
-            1 => Ok(AdminListActionType::Add),
-            2 => Ok(AdminListActionType::Remove),
-            3 => Ok(AdminListActionType::AddSuper),
-            4 => Ok(AdminListActionType::RemoveSuper),
-            _ => Err(IntentError::UnknownAdminListAction),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct UpdateAdminListIntentData {
-    pub action_type: AdminListActionType,
-    pub inbox_id: String,
-}
-
-impl UpdateAdminListIntentData {
-    pub fn new(action_type: AdminListActionType, inbox_id: String) -> Self {
-        Self {
-            action_type,
-            inbox_id,
-        }
-    }
-}
-
-impl From<UpdateAdminListIntentData> for Vec<u8> {
-    fn from(intent: UpdateAdminListIntentData) -> Self {
-        let mut buf = Vec::new();
-        let action_type = intent.action_type as i32;
-
-        UpdateAdminListsData {
-            version: Some(UpdateAdminListsVersion::V1(UpdateAdminListsV1 {
-                admin_list_update_type: action_type,
-                inbox_id: intent.inbox_id,
-            })),
-        }
-        .encode(&mut buf)
-        .expect("encode error");
-
-        buf
-    }
-}
-
-impl TryFrom<Vec<u8>> for UpdateAdminListIntentData {
-    type Error = IntentError;
-
-    fn try_from(data: Vec<u8>) -> Result<Self, Self::Error> {
-        let msg = UpdateAdminListsData::decode(Bytes::from(data))?;
-
-        let action_type: AdminListActionType = match msg.version {
-            Some(UpdateAdminListsVersion::V1(ref v1)) => {
-                AdminListActionType::try_from(v1.admin_list_update_type)?
-            }
-            None => return Err(IntentError::MissingUpdateAdminVersion),
-        };
-        let inbox_id = match msg.version {
-            Some(UpdateAdminListsVersion::V1(ref v1)) => v1.inbox_id.clone(),
-            None => return Err(IntentError::MissingUpdateAdminVersion),
-        };
-
-        Ok(Self::new(action_type, inbox_id))
-    }
-}
-
-#[repr(i32)]
-#[derive(Debug, Clone, PartialEq)]
-pub enum PermissionUpdateType {
-    AddMember = 1,      // Matches ADD_MEMBER in Protobuf
-    RemoveMember = 2,   // Matches REMOVE_MEMBER in Protobuf
-    AddAdmin = 3,       // Matches ADD_ADMIN in Protobuf
-    RemoveAdmin = 4,    // Matches REMOVE_ADMIN in Protobuf
-    UpdateMetadata = 5, // Matches UPDATE_METADATA in Protobuf
-}
-
-impl TryFrom<i32> for PermissionUpdateType {
-    type Error = IntentError;
-
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
-        match value {
-            1 => Ok(PermissionUpdateType::AddMember),
-            2 => Ok(PermissionUpdateType::RemoveMember),
-            3 => Ok(PermissionUpdateType::AddAdmin),
-            4 => Ok(PermissionUpdateType::RemoveAdmin),
-            5 => Ok(PermissionUpdateType::UpdateMetadata),
-            _ => Err(IntentError::UnknownPermissionUpdateType),
-        }
-    }
-}
-
-#[repr(i32)]
-#[derive(Debug, Clone, PartialEq)]
-pub enum PermissionPolicyOption {
-    Allow = 1,          // Matches ADD_MEMBER in Protobuf
-    Deny = 2,           // Matches REMOVE_MEMBER in Protobuf
-    AdminOnly = 3,      // Matches ADD_ADMIN in Protobuf
-    SuperAdminOnly = 4, // Matches REMOVE_ADMIN in Protobuf
-}
-
-impl TryFrom<i32> for PermissionPolicyOption {
-    type Error = IntentError;
-
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
-        match value {
-            1 => Ok(PermissionPolicyOption::Allow),
-            2 => Ok(PermissionPolicyOption::Deny),
-            3 => Ok(PermissionPolicyOption::AdminOnly),
-            4 => Ok(PermissionPolicyOption::SuperAdminOnly),
-            _ => Err(IntentError::UnknownPermissionPolicyOption),
-        }
-    }
-}
-
-impl From<PermissionPolicyOption> for MembershipPolicies {
-    fn from(value: PermissionPolicyOption) -> Self {
-        match value {
-            PermissionPolicyOption::Allow => MembershipPolicies::allow(),
-            PermissionPolicyOption::Deny => MembershipPolicies::deny(),
-            PermissionPolicyOption::AdminOnly => MembershipPolicies::allow_if_actor_admin(),
-            PermissionPolicyOption::SuperAdminOnly => {
-                MembershipPolicies::allow_if_actor_super_admin()
-            }
-        }
-    }
-}
-
-impl From<PermissionPolicyOption> for MetadataPolicies {
-    fn from(value: PermissionPolicyOption) -> Self {
-        match value {
-            PermissionPolicyOption::Allow => MetadataPolicies::allow(),
-            PermissionPolicyOption::Deny => MetadataPolicies::deny(),
-            PermissionPolicyOption::AdminOnly => MetadataPolicies::allow_if_actor_admin(),
-            PermissionPolicyOption::SuperAdminOnly => {
-                MetadataPolicies::allow_if_actor_super_admin()
-            }
-        }
-    }
-}
-
-impl From<PermissionPolicyOption> for PermissionsPolicies {
-    fn from(value: PermissionPolicyOption) -> Self {
-        match value {
-            PermissionPolicyOption::Allow => {
-                tracing::error!("PermissionPolicyOption::Allow is not allowed for PermissionsPolicies, set to super_admin only instead");
-                PermissionsPolicies::allow_if_actor_super_admin()
-            }
-            PermissionPolicyOption::Deny => PermissionsPolicies::deny(),
-            PermissionPolicyOption::AdminOnly => PermissionsPolicies::allow_if_actor_admin(),
-            PermissionPolicyOption::SuperAdminOnly => {
-                PermissionsPolicies::allow_if_actor_super_admin()
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct UpdatePermissionIntentData {
-    pub update_type: PermissionUpdateType,
-    pub policy_option: PermissionPolicyOption,
-    pub metadata_field_name: Option<String>,
-}
-
-impl UpdatePermissionIntentData {
-    pub fn new(
-        update_type: PermissionUpdateType,
-        policy_option: PermissionPolicyOption,
-        metadata_field_name: Option<String>,
-    ) -> Self {
-        Self {
-            update_type,
-            policy_option,
-            metadata_field_name,
-        }
-    }
-}
-
-impl From<UpdatePermissionIntentData> for Vec<u8> {
-    fn from(intent: UpdatePermissionIntentData) -> Self {
-        let mut buf = Vec::new();
-        let update_type = intent.update_type as i32;
-        let policy_option = intent.policy_option as i32;
-
-        UpdatePermissionData {
-            version: Some(UpdatePermissionVersion::V1(UpdatePermissionV1 {
-                permission_update_type: update_type,
-                permission_policy_option: policy_option,
-                metadata_field_name: intent.metadata_field_name,
-            })),
-        }
-        .encode(&mut buf)
-        .expect("encode error");
-
-        buf
-    }
-}
-
-impl TryFrom<Vec<u8>> for UpdatePermissionIntentData {
-    type Error = IntentError;
-
-    fn try_from(data: Vec<u8>) -> Result<Self, Self::Error> {
-        let msg = UpdatePermissionData::decode(Bytes::from(data))?;
-        let Some(UpdatePermissionVersion::V1(update_permission_data::V1 {
-            permission_update_type,
-            permission_policy_option,
-            metadata_field_name,
-        })) = msg.version
-        else {
-            return Err(IntentError::UnsupportedPermissionVersion);
-        };
-        let update_type: PermissionUpdateType = permission_update_type.try_into()?;
-        let policy_option: PermissionPolicyOption = permission_policy_option.try_into()?;
-        Ok(Self::new(update_type, policy_option, metadata_field_name))
     }
 }
