+++ xmtp_mls/src/groups/members.rs
@@ -1,4 +1,5 @@
-use super::{validated_commit::extract_group_membership, GroupError, MlsGroup, ScopedGroupClient};
+use super::{GroupError, MlsGroup, ScopedGroupClient};
+use crate::groups::mls_ext::MlsExtensionsExt;
 use xmtp_db::{
     association_state::StoredAssociationState,
     consent_record::{ConsentState, ConsentType},
@@ -40,7 +41,7 @@ where
         provider: &XmtpOpenMlsProvider,
     ) -> Result<Vec<GroupMember>, GroupError> {
         let group_membership = self.load_mls_group_with_lock(provider, |mls_group| {
-            Ok(extract_group_membership(mls_group.extensions())?)
+            Ok(mls_group.extensions().group_membership()?)
         })?;
         let requests = group_membership
             .members
