+++ xmtp_mls/src/groups/mls_ext.rs
@@ -1,3 +1,6 @@
 mod mls_ext_welcome;
 
 pub(crate) use mls_ext_welcome::*;
+
+mod intent;
+pub(crate) use intent::*;
