+++ xmtp_mls/src/groups/mod.rs
@@ -27,7 +27,6 @@ use self::{
         AdminListActionType, PermissionPolicyOption, PermissionUpdateType,
         UpdateAdminListIntentData, UpdateMetadataIntentData, UpdatePermissionIntentData,
     },
-    validated_commit::extract_group_membership,
 };
 use self::{
     group_metadata::{GroupMetadata, GroupMetadataError},
@@ -39,6 +38,7 @@ use crate::groups::group_mutable_metadata::{
     extract_group_mutable_metadata, MessageDisappearingSettings,
 };
 use crate::groups::intents::UpdateGroupMembershipResult;
+use crate::groups::mls_ext::MlsExtensionsExt;
 use crate::GroupCommitLock;
 use crate::{
     client::{ClientError, XmtpMlsLocalContext},
@@ -56,7 +56,7 @@ use crate::{
 };
 use device_sync::preference_sync::UserPreferenceUpdate;
 use intents::SendMessageIntentData;
-use mls_ext::DecryptedWelcome;
+use mls_ext::{DecryptedWelcome, GroupExtError};
 use mls_sync::GroupMessageProcessingError;
 use openmls::{
     credentials::CredentialType,
@@ -224,6 +224,8 @@ pub enum GroupError {
     TooManyCharacters { length: usize },
     #[error("Group is paused until version {0} is available")]
     GroupPausedUntilUpdate(String),
+    #[error(transparent)]
+    GroupExtError(#[from] GroupExtError),
 }
 
 impl RetryableError for GroupError {
@@ -232,6 +234,7 @@ impl RetryableError for GroupError {
             Self::ReceiveErrors(errors) => errors.iter().any(|e| e.is_retryable()),
             Self::Client(client_error) => client_error.is_retryable(),
             Self::Storage(storage) => storage.is_retryable(),
+            Self::GroupExtError(g) => g.is_retryable(),
             Self::ReceiveError(msg) => msg.is_retryable(),
             Self::Hpke(hpke) => hpke.is_retryable(),
             Self::Identity(identity) => identity.is_retryable(),
@@ -2000,7 +2003,7 @@ async fn validate_initial_group_membership(
 ) -> Result<(), GroupError> {
     tracing::info!("Validating initial group membership");
     let extensions = staged_welcome.public_group().group_context().extensions();
-    let membership = extract_group_membership(extensions)?;
+    let membership = extensions.group_membership()?;
     let needs_update = conn.filter_inbox_ids_needing_updates(membership.to_filters().as_slice())?;
     if !needs_update.is_empty() {
         let ids = needs_update.iter().map(AsRef::as_ref).collect::<Vec<_>>();
